Add files via upload
This project is the final project from Hacktiv8 course (surely not the last project from me). The goal of this project is to showcase all the knowledge in data science (starting from uploading data, data pre processing to machine learning model). For this project I choose Kaggle Competition of Titanic- Machine Learning from Disaster. The challenge is to predict the likelihood of a passenger survival based on the data we use to train the machine learning model.